filtering by the first letter of the mineral name
Add links for each letter of the alphabet.
This should be added to the layout template so that it appears on every page.
When a letter is clicked, a list of minerals that start with that letter should be displayed in the list view.
The letter of the alphabet currently being displayed should be bolded. In the details view, no letter should be bolded.
On the homepage, select ‘A’ by default.